se añadio escript sql al repositorio

+++ GestorComunicaciones.WebApi/Controllers/Communications/CommunicationsController.cs
@@ -8,10 +8,12 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -24,16 +26,20 @@ namespace GestorComunicaciones.WebApi.Controllers.Communications
     public class CommunicationsController : ControllerBase
     {
         private readonly IMapper _mapper;
+        private readonly ILogger _logger;
         private readonly ICommunicationService _communicationService;
-        public CommunicationsController(IMapper mapper, ICommunicationService communicationService)
+        public CommunicationsController(IMapper mapper,ILogger logger, ICommunicationService communicationService)
         {
             _mapper = mapper;
+            _logger = logger;
             _communicationService = communicationService;
         }
         // GET: api/<CommunicationsController>
         [HttpGet]
         public async Task<IEnumerable<CommunicationDto>> GetAsync()
         {
+            string email = User.FindFirstValue(ClaimTypes.Email);
+            _logger.LogInformation("usuario{usuario}esta consultando las comunicaciones",email);
             var communications = await this._communicationService.GetCommunicationsAsync();
             var communicationsDto = _mapper.Map<IEnumerable<CommunicationDto>>(communications);
             return communicationsDto;
@@ -50,27 +56,54 @@ namespace GestorComunicaciones.WebApi.Controllers.Communications
 
         // POST api/<CommunicationsController>
         [HttpPost]
-        public async Task Post([FromBody] CommunicationDto communicationDto)
+        public async Task<IActionResult> Post([FromBody] CommunicationDto communicationDto)
         {
-            var communication = _mapper.Map<Communication>(communicationDto);
-            await this._communicationService.InsertCommunication(communication);
+            string email = User.FindFirstValue(ClaimTypes.Email);
+            _logger.LogInformation("usuario{usuario}esta intentando insertar comunicacion", email);
+            string Role = User.FindFirstValue(ClaimTypes.Role);
+            if(Role.ToLower()=="administrador"|| Role.ToLower() == "gestor")
+            {
+                var communication = _mapper.Map<Communication>(communicationDto);
+                await this._communicationService.InsertCommunication(communication);
+
+                return Ok(new { succes = true });
+                
+            }
+            return Unauthorized(new { succes = false, msg = "solo el administrador o el gestor pueden crear comunicaciones" });
+
         }
 
         // PUT api/<CommunicationsController>/5
         [HttpPut("{id}")]
         public async Task<IActionResult> PutAsync(long id, [FromBody] CommunicationDto communicationDto)
         {
-            var communication = _mapper.Map<Communication>(communicationDto);
-            var result = await this._communicationService.UpdateCommunication(communication);
-            return Ok(result);
+            string email = User.FindFirstValue(ClaimTypes.Email);
+            _logger.LogInformation("usuario{usuario}esta intentando actualizar comunicacion", email);
+            string Role = User.FindFirstValue(ClaimTypes.Role);
+            if (Role.ToLower() == "administrador" || Role.ToLower() == "gestor")
+            {
+                var communication = _mapper.Map<Communication>(communicationDto);
+                var result = await this._communicationService.UpdateCommunication(communication);
+                return Ok(result);
+            }
+            return Unauthorized(new { succes = false, msg = "solo el administrador o el gestor pueden editar comunicaciones" });
         }
 
         // DELETE api/<CommunicationsController>/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteAsync(long id)
         {
-            var result = await this._communicationService.DeleteCommunication(id);
-            return Ok(result);
+            string email = User.FindFirstValue(ClaimTypes.Email);
+            _logger.LogInformation("usuario{usuario}esta intentando borrar comunicacion", email);
+            string Role = User.FindFirstValue(ClaimTypes.Role);
+            if (Role.ToLower() == "administrador" || Role.ToLower() == "gestor")
+            {
+                var result = await this._communicationService.DeleteCommunication(id);
+                return Ok(result);                
+
+            }
+            return Unauthorized(new { succes = false, msg = "solo el administrador o el gestor pueden borrar comunicaciones" });
+
         }
     }
 }
